add delte confirm modal, delete backend auth, auth guard

+++ server/api/poll/poll.service.js
@@ -1,6 +1,5 @@
 const dbService = require('../../services/db.service')
 const ObjectId = require('mongodb').ObjectId
-const asyncLocalStorage = require('../../services/als.service')
 
 async function query(filterBy = {}) {
     try {
@@ -54,16 +53,16 @@ async function add(poll) {
     }
 }
 
-async function remove(pollId) {
+async function remove(pollId, user) {
     try {
-        // const store = asyncLocalStorage.getStore()
-        // const { userId, isAdmin } = store
+
         const collection = await dbService.getCollection('poll')
         // remove only if user is owner/admin
         const query = { _id: ObjectId(pollId) }
-        // if (!isAdmin) query.byUserId = ObjectId(userId)
+        const poll = await collection.findOne({ '_id': ObjectId(pollId) })
+        console.log('THE POLL IS ', poll)
+        if(poll.owner._id !== user._id) throw new Error('Not the owner')
         return await collection.deleteOne(query)
-        // return await collection.deleteOne({ _id: ObjectId(pollId), byUserId: ObjectId(userId) })
     } catch (err) {
         logger.error(`cannot remove review ${pollId}`, err)
         throw err
